fix(react): split EVM and Solana into separate subpath exports
Round 5 codex caught: even with the optional peer dep architecture,
importing @stwd/react/wallet eagerly evaluated BOTH the EVM and
Solana provider files because they shared a single source file.
EVM-only consumers who installed only EVM peers would crash with
ERR_MODULE_NOT_FOUND on the Solana adapter imports.

Fix: split providers/WalletProviders.tsx into two siblings:
  - providers/EVMProvider.tsx   (only EVM peer imports)
  - providers/SolanaProvider.tsx (only Solana peer imports)

Add new package.json exports map entries:
  - @stwd/react/wallet/evm    -> dist/wallet/evm.js
  - @stwd/react/wallet/solana -> dist/wallet/solana.js

The combined @stwd/react/wallet entry remains for apps using both
chains. Existing imports continue to work via the barrel re-export.

Verified clean SSR import:
  $ node --input-type=module -e "import('./packages/react/dist/wallet/evm.js')..."
  evm OK: EVMWalletProvider,createDefaultWagmiConfig
  $ node --input-type=module -e "import('./packages/react/dist/wallet/solana.js')..."
  solana OK: DEFAULT_SOLANA_WALLETS,SolanaWalletProvider,createDefaultSolanaWallets

README updated with per-chain entry guidance.
