+++ packages/react/src/providers/WalletProviders.tsx
@@ -1,240 +1,23 @@
-import {
-  connectorsForWallets,
-  darkTheme,
-  RainbowKitProvider,
-  type Theme,
-} from "@rainbow-me/rainbowkit";
-import {
-  coinbaseWallet,
-  injectedWallet,
-  ledgerWallet,
-  metaMaskWallet,
-  phantomWallet,
-  rabbyWallet,
-  rainbowWallet,
-  safeWallet,
-  trustWallet,
-  walletConnectWallet,
-} from "@rainbow-me/rainbowkit/wallets";
-// Import each Solana adapter from its own subpath rather than from the
-// barrel `@solana/wallet-adapter-wallets`. The barrel re-exports every
-// adapter (including @solana/wallet-adapter-ledger and -trezor) so any
-// import from it forces Node ESM to resolve the entire dependency tree,
-// which fails with ERR_MODULE_NOT_FOUND because @ledgerhq/errors ships
-// extension-less imports incompatible with strict ESM resolution. By
-// going directly to subpackages we keep the SSR/Node import path clean.
-import { Coin98WalletAdapter } from "@solana/wallet-adapter-coin98";
-import { CoinbaseWalletAdapter } from "@solana/wallet-adapter-coinbase";
-import { MathWalletAdapter } from "@solana/wallet-adapter-mathwallet";
-import { PhantomWalletAdapter } from "@solana/wallet-adapter-phantom";
-import {
-  ConnectionProvider,
-  WalletProvider,
-  type WalletProviderProps,
-} from "@solana/wallet-adapter-react";
-import { WalletModalProvider } from "@solana/wallet-adapter-react-ui";
-import { SolflareWalletAdapter } from "@solana/wallet-adapter-solflare";
-import { TrustWalletAdapter } from "@solana/wallet-adapter-trust";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { type ReactNode, useMemo } from "react";
-import type { Chain, Transport } from "viem";
-// All imports from optional peer dependencies are intentional. These two
-// wrappers are opt-in utilities; consumers who ship their own wagmi / Solana
-// providers should import them directly instead of using these.
-import { createConfig, http, type Config as WagmiConfig, WagmiProvider } from "wagmi";
-
-// ─── EVM wrapper ─────────────────────────────────────────────────────────────
-
-export type DefaultWagmiChains = readonly [Chain, ...Chain[]];
-
-type DefaultWagmiTransports<TChains extends DefaultWagmiChains> = Record<
-  TChains[number]["id"],
-  Transport
->;
-
-export interface CreateDefaultWagmiConfigOptions<TChains extends DefaultWagmiChains> {
-  /** WalletConnect Cloud project ID. Get one at https://cloud.walletconnect.com. */
-  projectId: string;
-  /** wagmi chains to support. Must include at least one chain. */
-  chains: TChains;
-  /** App name shown in wallet connection prompts. Defaults to "Steward". */
-  appName?: string;
-  /** Enable wagmi SSR support. Defaults to true. */
-  ssr?: boolean;
-}
-
-function createDefaultTransports<TChains extends DefaultWagmiChains>(
-  chains: TChains,
-): DefaultWagmiTransports<TChains> {
-  return Object.fromEntries(
-    chains.map((chain) => [chain.id, http()]),
-  ) as DefaultWagmiTransports<TChains>;
-}
-
 /**
- * Creates a wagmi config with Steward's curated RainbowKit wallet order.
- * Consumers can keep passing their own config to EVMWalletProvider when they
- * need full control.
- */
-export function createDefaultWagmiConfig<TChains extends DefaultWagmiChains>({
-  projectId,
-  chains,
-  appName = "Steward",
-  ssr = true,
-}: CreateDefaultWagmiConfigOptions<TChains>): WagmiConfig {
-  const connectors = connectorsForWallets(
-    [
-      {
-        groupName: "Recommended",
-        wallets: [
-          metaMaskWallet,
-          coinbaseWallet,
-          walletConnectWallet,
-          rainbowWallet,
-          rabbyWallet,
-          trustWallet,
-          phantomWallet,
-          ledgerWallet,
-          safeWallet,
-          injectedWallet,
-        ],
-      },
-    ],
-    { appName, projectId },
-  );
-
-  return createConfig({
-    chains,
-    connectors,
-    transports: createDefaultTransports(chains),
-    ssr,
-  }) as WagmiConfig;
-}
-
-export interface EVMWalletProviderProps {
-  /** wagmi v2 `Config` created with `createConfig()` or `createDefaultWagmiConfig()`. */
-  config: WagmiConfig;
-  /**
-   * TanStack Query client. Pass yours if the host app already has one;
-   * otherwise a default client is created and scoped to this subtree.
-   */
-  queryClient?: QueryClient;
-  /** RainbowKit theme. Defaults to dark. Pass `null` to skip theming. */
-  theme?: Theme | null;
-  /** RainbowKit modal size. Defaults to "compact". */
-  modalSize?: "compact" | "wide";
-  /** Reconnect on mount. Defaults to true. */
-  reconnectOnMount?: boolean;
-  children?: ReactNode;
-}
-
-/**
- * Wraps children with wagmi + RainbowKit + TanStack Query providers. This is
- * an optional convenience. Most apps already have their own wagmi setup;
- * skip this wrapper if so. `<WalletLogin chains="evm">` only needs the
- * ambient wagmi + RainbowKit + QueryClient context to exist above it.
+ * Backwards-compatible barrel that re-exports both EVM and Solana
+ * provider modules. Importing this file evaluates both peer-dep trees,
+ * so consumers who only need one chain should import from
+ * `./EVMProvider.js` or `./SolanaProvider.js` directly (or use the
+ * `@stwd/react/wallet/evm` and `@stwd/react/wallet/solana` subpath
+ * exports).
  *
- * Remember to import `@rainbow-me/rainbowkit/styles.css` once at your app root.
+ * Existing imports from `@stwd/react/wallet` continue to work.
  */
-export function EVMWalletProvider({
-  config,
-  queryClient,
-  theme,
-  modalSize = "compact",
-  reconnectOnMount = true,
-  children,
-}: EVMWalletProviderProps) {
-  const resolvedTheme = theme === null ? null : (theme ?? darkTheme());
-  const client = useMemo(() => queryClient ?? new QueryClient(), [queryClient]);
-  return (
-    <WagmiProvider config={config} reconnectOnMount={reconnectOnMount}>
-      <QueryClientProvider client={client}>
-        <RainbowKitProvider theme={resolvedTheme} modalSize={modalSize}>
-          {children}
-        </RainbowKitProvider>
-      </QueryClientProvider>
-    </WagmiProvider>
-  );
-}
 
-// ─── Solana wrapper ──────────────────────────────────────────────────────────
-
-// Default Solana wallet adapters. Hardware wallets (Ledger, Trezor) are
-// intentionally excluded from defaults: their adapter packages do not
-// implement `signMessage`, so they cannot complete the SIWS sign-in flow.
-// Additionally, `@ledgerhq/errors` ships a non-extension import that
-// `ERR_MODULE_NOT_FOUND`s under Node ESM/SSR, which would break apps that
-// import `@stwd/react/wallet` server-side. Apps that want hardware support
-// in non-login contexts can extend the wallet list explicitly.
-//
-// Factory function: each call returns FRESH adapter instances. Solana
-// wallet adapters keep mutable connection state and event listeners on
-// the instance, so sharing one set across multiple <SolanaWalletProvider>
-// mounts would leak state/events between providers. Always call this
-// from a useMemo inside the provider so each provider mount gets its
-// own adapters.
-export function createDefaultSolanaWallets(): WalletProviderProps["wallets"] {
-  return [
-    new PhantomWalletAdapter(),
-    new SolflareWalletAdapter(),
-    new CoinbaseWalletAdapter(),
-    new TrustWalletAdapter(),
-    new MathWalletAdapter(),
-    new Coin98WalletAdapter(),
-  ];
-}
-
-/**
- * @deprecated Module-level array kept for backwards compatibility with
- * consumers who imported `DEFAULT_SOLANA_WALLETS` directly. Adapter
- * instances in this array are SHARED across all consumers, so passing
- * this to multiple <SolanaWalletProvider> mounts will leak connection
- * state. Prefer calling `createDefaultSolanaWallets()` per provider, or
- * just leave `<SolanaWalletProvider>`'s `wallets` prop unset (the
- * provider builds fresh adapters internally).
- */
-export const DEFAULT_SOLANA_WALLETS =
-  createDefaultSolanaWallets() satisfies WalletProviderProps["wallets"];
-
-export interface SolanaWalletProviderProps {
-  /** JSON-RPC endpoint (`https://api.mainnet-beta.solana.com`, Helius, etc.). */
-  endpoint: string;
-  /**
-   * Wallet adapters. Defaults to DEFAULT_SOLANA_WALLETS. Pass an explicit
-   * array to narrow, replace, or extend the list.
-   *
-   * Backpack, Brave, and other Solana Wallet Standard wallets are discovered
-   * by wallet-adapter at runtime when the browser wallet is present.
-   */
-  wallets?: WalletProviderProps["wallets"];
-  /** Auto-connect previously selected wallet on mount. Defaults to true. */
-  autoConnect?: boolean;
-  children?: ReactNode;
-}
-
-/**
- * Wraps children with Solana wallet-adapter providers. Optional convenience.
- *
- * Remember to import `@solana/wallet-adapter-react-ui/styles.css` once at your
- * app root.
- */
-export function SolanaWalletProvider({
-  endpoint,
-  wallets,
-  autoConnect = true,
-  children,
-}: SolanaWalletProviderProps) {
-  // Build a fresh adapter list per provider mount when no wallets prop
-  // was passed. Sharing a single module-level array across mounts would
-  // leak connection state between providers (the previous behavior was
-  // a P2 codex finding).
-  const resolvedWallets = useMemo(() => wallets ?? createDefaultSolanaWallets(), [wallets]);
-
-  return (
-    <ConnectionProvider endpoint={endpoint}>
-      <WalletProvider wallets={resolvedWallets} autoConnect={autoConnect}>
-        <WalletModalProvider>{children}</WalletModalProvider>
-      </WalletProvider>
-    </ConnectionProvider>
-  );
-}
+export type {
+  CreateDefaultWagmiConfigOptions,
+  DefaultWagmiChains,
+  EVMWalletProviderProps,
+} from "./EVMProvider.js";
+export { createDefaultWagmiConfig, EVMWalletProvider } from "./EVMProvider.js";
+export type { SolanaWalletProviderProps } from "./SolanaProvider.js";
+export {
+  createDefaultSolanaWallets,
+  DEFAULT_SOLANA_WALLETS,
+  SolanaWalletProvider,
+} from "./SolanaProvider.js";
